fix(App): Support taking proxy manager via props
